requirements.txt
numpy==1.26.4
opencv_python==4.9.0.80
Pillow==10.3.0
scikit_learn==1.4.2
timm==0.4.5
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.2
